feat/todo-content
# features
- Refactoring and code cleaning
